[34535] Override project limiting filter if subproject filter present
When the query has an explicit SubProject or OnlySubproject filter, the
project limiting filter based on the global settting does not really
make any sense. It will prevent seeing any results when trying to view
work packages from a subproject unless it is enabled globally.

This change suggests to skip this limiting filter whenever an explicit
subproject filter is added to the query.

This fixes subproject boards when the global setting is enabled, since
that sets only the `onlySubprojectFilter`.

https://community.openproject.com/wp/34535