[34490] Don't try to redirect angular route on first load
When we click on a ui router state link on a page that is not routed by
angular, the route has to be cancelled and we need to move to the actual
URL. Otherwise, ui-router will simply update the URL but due to the
missing ui-view in the DOM, nothing will happen.

This check does not properly take into account the initial page load, in
which case we should obviously not redirect.

https://community.openproject.com/wp/34490